chore: refactor cmds
