+++ internal/quickfs/quickfs.go
@@ -11,6 +11,8 @@ import (
 	"github.com/rs/zerolog/log"
 )
 
+var CountryCodes = []string{"US", "AT", "AU", "BE", "CA", "CH", "DE", "DK", "ES", "FI", "FI", "FR", "GR", "IT", "LN", "MM", "NL", "NO", "NZ", "PL", "SE"}
+
 type Data struct {
 	Shares       int     `json:"shares"`
 	TaxRate      float64 `json:"taxRate"`
